+++ R/test_ejam.R
@@ -342,6 +342,7 @@ instead of tests/testthat/_logs
         "test-ejamit_sitetype_from_output.R",
 
         "test-ejam2report.R",
+        "test-build_community_report.R",
         "test-ejam2excel.R",
         "test-ejam2barplot_sites.R",
         "test-ejam2barplot_indicators.R",
@@ -602,7 +603,8 @@ and all filenames listed there actually exist as in that folder called `test`.\n
               "test-naics_is.valid.R", "test-naics_subcodes_from_code.R", "test-test1.R",
               "test-test2.R", "test-URL_FUNCTIONS_part1.R", "test-URL_FUNCTIONS_part2.R",
               "test-ejamapi.R", "test-url_columns_bysite.R", "test-url_ejamapi.R",
-              "test-url_package.R", "test-ejam2boxplot_ratios.R", "test-shiny-1-14-compat.R"),
+              "test-url_package.R", "test-ejam2boxplot_ratios.R", "test-shiny-1-14-compat.R",
+              "test-build_community_report.R"),
           seconds_byfile =
             c(0, 0, 0, 0, 0, 0, 0,
               0, 0, 2, 0, 1, 0, 1, 0, 0, 0, 2, 0, 0, 0, 0, 1, 0, 6, 0, 0, 0,
@@ -610,9 +612,9 @@ and all filenames listed there actually exist as in that folder called `test`.\n
               39, 0, 11, 11, 0, 5, 32, 5, 0, 5, 0, 2, 0, 140, 2, 8, 0, 3, 2,
               23, 0, 7, 0, 12, 0, 4, 5, 1, 0, 3, 0, 3, 0, 0, 0, 0, 1, 0, 4,
               2, 0, 0, 0, 0, 1, 0, 0, 16, 8, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0,
-              0, 0, 0, 2, 1, 0, 0, 0, 0, 0, 0, 0, 12, 0, 4, 0, 0, 0, 0)),
+              0, 0, 0, 2, 1, 0, 0, 0, 0, 0, 0, 0, 12, 0, 4, 0, 0, 0, 0, 1)),
           row.names = c(NA,
-                        -126L), class = "data.frame")
+                        -127L), class = "data.frame")
       )